+++ src/components/declarations/dossier/files-details.js
