+++ src/components/declarations/file-validation-errors.js
