+++ src/components/declarations/prelevement-type-badge.js
