+++ src/components/declarations/type-saisie-badge.js
