+++ src/components/prelevements/point-exploitations.js
@@ -17,7 +17,7 @@ import {orderBy} from 'lodash-es'
 
 import {getVolumesExploitation} from '@/app/api/points-prelevement.js'
 import Document from '@/components/document.js'
-import VolumesChart from '@/components/points-prelevement/volumes-chart.js'
+import VolumesChart from '@/components/prelevements/volumes-chart.js'
 import Regles from '@/components/regles.js'
 import formatDate from '@/lib/format-date.js'
